add SimdOperations for vector but not tested and not finished!!

+++ PP/MatrixWoutSIMD/Element.cs
@@ -4,9 +4,9 @@
     {
         public float Value { get; set; }
 
-        public float Row { get; set; }
+        public int Row { get; set; }
 
-        public float Column { get; set; }
+        public int Column { get; set; }
 
         public override string ToString()
         {
